split supabaseApi into .api.ts and .models.ts
