+++ src/supabase/supabase.api.ts
@@ -1,4 +1,5 @@
 import { supabase } from "./supabase";
+import { Latecoming, Profile } from "./supabase.models";
 
 export const getProfiles = async (): Promise<Profile[]> => {
   const { data, error } = await supabase.from("profiles").select("*");
@@ -30,21 +31,3 @@ export const addNewLatecoming = async (uuid: string, minutes: number) => {
 
   if (error) throw error;
 };
-
-export interface Latecoming {
-  id: number;
-  created_at: Date;
-  guilty: string;
-  minutes: number;
-}
-
-export interface Profile {
-  id: string;
-  username: string;
-  ppu: number;
-}
-
-export interface LatecomingProfile extends Profile {
-  totalMinutes: number;
-  earnedPpu?: number;
-}
